Move decimal parser from ValueFormatter to SendAmount module
